# web/openapi/raffa-api.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Raffa API",
    "version": "1.1.0",
...
  },
  "servers": [
    {
      "url": "/",
      "description": "Relative to the runtime-configured API origin (AppConfig.apiBaseUrl, ADR-012 \"config, not code\"; see web/src/config/appConfig.ts). This document intentionally does not hard-code a dev/demo host."
    }
  ],
  "paths": {
...
    "/api/savings": {
      "get": {
        "operationId": "getSavingsOpportunities",
        "summary": "Home screen's opportunities table: every SavingsOpportunity for the caller's tenant, newest identified first.",
        "description": "Maps to GetSavingsOpportunitiesAsync in backend/src/Raffa.Api/SavingsEndpointExtensions.cs (task E04/F02/US02/T01 savings-opportunity, task E04/F03/US01/T02 savings-list AC-3 confidenceLevel). Task E08/F02/US01/T01 (savings-home, ADR-020 screen 9) adds this operation's GET half -- the fifth web epic to extend this document (see this document's own info.description 'repeating chore' provenance paragraph); PATCH /api/savings/{id} (status/owner/realized-value edits) already exists on the backend but has no screen in this task's own AC that needs it, so it stays undocumented here until a future editing surface has a real caller (the same 'wiring lands with the first real caller' convention this codebase already follows elsewhere). Never 404s -- an empty items array is a tenant's own honest, valid 'no opportunities yet' answer. confidenceLevel/status are real, closed three-value enums (SavingsConfidenceLevel.ToString() / SavingsOpportunityStatus.ToString()), declared with enum the same way GET /api/renewals's own status field already is. type is free text (Domain.SavingsOpportunity.Type's own doc comment: no ADR/spec fixes a vocabulary for it).",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tenant-scoped opportunity list, newest identified first.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "totalCount"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "supplierId",
                          "contractId",
                          "type",
                          "currentSpend",
                          "currency",
                          "estimatedSavingsLow",
                          "estimatedSavingsHigh",
                          "confidence",
                          "confidenceLevel",
                          "status",
                          "owner",
                          "createdAt",
                          "updatedAt",
                          "realizedAmount"
                        ],
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "supplierId": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "format": "uuid"
                          },
                          "contractId": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "format": "uuid"
                          },
                          "type": {
                            "type": "string"
                          },
                          "currentSpend": {
                            "type": "number"
                          },
                          "currency": {
                            "type": "string"
                          },
                          "estimatedSavingsLow": {
                            "type": "number"
                          },
                          "estimatedSavingsHigh": {
                            "type": "number"
                          },
                          "confidence": {
                            "type": "number"
                          },
                          "confidenceLevel": {
                            "type": "string",
                            "enum": [
                              "Low",
                              "Medium",
                              "High"
                            ]
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "Identified",
                              "InProgress",
                              "Realized"
                            ]
                          },
                          "owner": {
                            "type": [
                              "string",
                              "null"
                            ]
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "realizedAmount": {
                            "type": [
                              "number",
                              "null"
                            ]
                          }
                        }
                      }
                    },
                    "totalCount": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing/invalid X-Tenant-Id header.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  }
}